(Board Integration) Integrated with background styling.
