+++ src/components/Application/content/Board.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
+import { useSelector } from 'react-redux';
 
 const BoardWrapper = styled.div`
   flex-grow: 9;
@@ -10,6 +11,21 @@ const BoardWrapper = styled.div`
   border-width: 5px;
 `;
 
-const Board = () => <BoardWrapper>Board</BoardWrapper>;
+const Board = () => {
+  const level = useSelector(state => state.game.level);
+  const { TopRow, BottomRow } = level;
 
+  const TopComponent = TopRow || (() => <div />);
+  const BottomComponent = BottomRow || (() => <div />);
+  return (
+    <BoardWrapper>
+      {level ? (
+        <>
+          <TopComponent />
+          <BottomComponent />
+        </>
+      ) : null}
+    </BoardWrapper>
+  );
+};
 export default Board;
