+++ src/components/Application/index.js
@@ -1,9 +1,12 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import Board from '../Board';
+// import Board from '../Board';
+import { useDispatch } from 'react-redux';
 import Content from './content/Content';
 import SideBar from './sidebar/SideBar';
+import { startGameAction } from '../../reducers/gameReducer';
+import LEVELS from '../LEVELS';
 
 const Layout = styled.div`
   display: flex;
@@ -15,12 +18,17 @@ const Layout = styled.div`
 
 const Application = ({ id = 1337 }) => {
   // id of a joined room
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(startGameAction(LEVELS.level1));
+    console.log('fuck');
+  }, []);
   if (!id) return <div>Should return to Main Screen</div>;
   return (
     <Layout>
       <SideBar />
       <Content />
-      <Board />
+      {/* <Board /> */}
     </Layout>
   );
 };
